Use full copy of powerline bindings directory

+++ app-misc/powerline/powerline-9999-r18.ebuild
@@ -56,6 +56,7 @@ RDEPEND="
 
 # Source directory from which all applicable files will be installed.
 POWERLINE_BINDINGS_DIR="${T}/bindings"
+POWERLINE_FULL_BINDINGS_DIR="${T}/bindings-full"
 
 # Target directory to which all applicable files will be installed.
 POWERLINE_TRG_DIR='/usr/share/powerline'
@@ -101,6 +102,9 @@ python_prepare_all() {
 	# distutils-based installation, copy them to such location and then remove
 	# them from the original tree that distutils operates on.
 	cp -R "${S}"/powerline/bindings "${POWERLINE_BINDINGS_DIR}" || die '"cp" failed.'
+	if use test ; then
+		cp -R "${S}"/powerline/bindings "${POWERLINE_FULL_BINDINGS_DIR}" || die '"cp" failed.'
+	fi
 
 	# Remove all non-Python files from the original tree.
 	find "${S}"/powerline/bindings -type f -not -name '*.py' -delete
@@ -140,7 +144,7 @@ python_compile_all() {
 python_test() {
 	# *All* bindings files are required for tests.
 	mv "${S}"/powerline/bindings{,.bak}
-	cp -R "${POWERLINE_BINDINGS_DIR}" "${S}"/powerline/bindings
+	cp -R "${POWERLINE_FULL_BINDINGS_DIR}" "${S}"/powerline/bindings
 	# Powerline shell tests do not work with LD_PRELOAD-based sandbox.
 	env -i LANG=en_US.UTF-8 PATH="$PATH" PYTHON="${PYTHON}" "${S}"/tests/test.sh || die 'Unit tests failed.'
 	rm -r "${S}"/powerline/bindings
